+++ src/ChuckNorrisJokesServiceProvider.php
@@ -2,8 +2,10 @@
 
 namespace Ejdelmonico\ChuckNorrisJokes;
 
+use Illuminate\Support\Facades\Route;
 use Illuminate\Support\ServiceProvider;
 use Ejdelmonico\ChuckNorrisJokes\Console\ChuckNorrisJoke;
+use Ejdelmonico\ChuckNorrisJokes\Http\Controllers\ChuckNorrisController;
 
 class ChuckNorrisJokesServiceProvider extends ServiceProvider
 {
@@ -14,6 +16,16 @@ class ChuckNorrisJokesServiceProvider extends ServiceProvider
                 ChuckNorrisJoke::class,
             ]);
         }
+
+        $this->loadViewsFrom(__DIR__ . '/../resources/views', 'chuck-norris');
+        $this->publishes([
+            __DIR__ . '/../resources/views' => resource_path('views/vendor/chuck-norris'),
+        ], 'views');
+        $this->publishes([
+            __DIR__ . '/../config/chuck-norris.php' => base_path('config/chuck-norris.php'),
+        ], 'config');
+
+        Route::get(config('chuck-norris.route'), ChuckNorrisController::class);
     }
 
     public function register()
@@ -21,5 +33,7 @@ class ChuckNorrisJokesServiceProvider extends ServiceProvider
         $this->app->bind('chuck-norris', function () {
             return new JokeFactory();
         });
+
+        $this->mergeConfigFrom(__DIR__ . '/../config/chuck-norris.php', 'chuck-norris');
     }
 }