+++ src/Http/Controllers/ChuckNorrisController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Ejdelmonico\ChuckNorrisJokes\Http\Controllers;
+
+use Ejdelmonico\ChuckNorrisJokes\Facades\ChuckNorris;
+
+class ChuckNorrisController
+{
+    public function __invoke()
+    {
+        return view('chuck-norris::joke', [
+            'joke' => ChuckNorris::getRandomJoke(),
+        ]);
+    }
+}