+++ tests/LaravelTest.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace Ejdelmonico\ChuckNorrisJokes\Tests;
+
+use Orchestra\Testbench\TestCase;
+use Illuminate\Support\Facades\Artisan;
+use Ejdelmonico\ChuckNorrisJokes\Facades\ChuckNorris;
+use Ejdelmonico\ChuckNorrisJokes\Console\ChuckNorrisJoke;
+use Ejdelmonico\ChuckNorrisJokes\ChuckNorrisJokesServiceProvider;
+
+class LaravelTest extends TestCase
+{
+    protected function getPackageProviders($app)
+    {
+        return [ChuckNorrisJokesServiceProvider::class];
+    }
+
+    protected function getPackageAliases($app)
+    {
+        return ['ChuckNorris' => ChuckNorrisJoke::class];
+    }
+
+    /** @test */
+    public function the_console_command_returns_a_joke()
+    {
+        $this->withoutMockingConsoleOutput();
+
+        ChuckNorris::shouldReceive('getRandomJoke')
+            ->once()
+            ->andReturn('some joke');
+
+        $this->artisan('chuck-norris');
+
+        $output = Artisan::output();
+
+        $this->assertSame('some joke' . PHP_EOL, $output);
+    }
+
+    /** @test */
+    public function the_route_can_be_accessed()
+    {
+        ChuckNorris::shouldReceive('getRandomJoke')
+            ->once()
+            ->andReturn('some joke');
+
+        $this->get('/chuck-norris')
+            ->assertViewIs('chuck-norris::joke')
+            ->assertViewHas('joke', 'some joke')
+            ->assertStatus(200);
+    }
+}