+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2019-05-27
+
+### Added
+
+- Laravel service provider
+- Laravel facade
+- Laravel package auto discovery
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
 ## [1.0.0] - 2019-05-26
 
 - Initial release